Fixes https://github.com/wso2/product-iots/issues/773
