--- modules/distribution/src/core/samples/sampledevice/component/api/src/main/java/org/wso2/carbon/sampledevice/api/util/ZipArchive.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.sampledevice.api.util;
-
-
-import java.io.File;
-
-/**
- * This is an utility class to hold zip files.
- */
-public class ZipArchive {
-
-    private File zipFile = null;
-    private String fileName = null;
-
-    public ZipArchive(String fileName, File zipFile) {
-        this.fileName = fileName;
-        this.zipFile = zipFile;
-    }
-
-    public File getZipFile() {
-        return zipFile;
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-}
